Problem: Large file upload consumes large memory (100MB file upload requires 100MB+ memory).
The cause of this issue is that Multipart::Parser uses StringScanner as buffer and appends chunk here,
```
      def on_read content
        handle_empty_content!(content)
        @sbuf.concat content #<- here
        run_parser
      end
```
and it keeps `@sbuf` growing but never discards processed chunk.
This PR discards chunk in buffer that have already processed in order to prevent huge memory consumption.