@tainoe 2019-06-18
I borrowed benchmark code from this :
https://gist.github.com/tompng/54eb216d81c600a46fef28ddab68b356
which is used in the PR :
https://github.com/rack/rack/pull/1193
My result is as follows.
```
#before
{"file_size"=>1000000, "run_parser_called"=>1, "memory"=>"12MB"}: 0.227233574sec
{"file_size"=>2000000, "run_parser_called"=>2, "memory"=>"14MB"}: 0.180755913sec
{"file_size"=>4000000, "run_parser_called"=>4, "memory"=>"17MB"}: 0.166251455sec
{"file_size"=>8000000, "run_parser_called"=>8, "memory"=>"23MB"}: 0.224019834sec
{"file_size"=>16000000, "run_parser_called"=>16, "memory"=>"36MB"}: 0.333338561sec
{"file_size"=>32000000, "run_parser_called"=>31, "memory"=>"61MB"}: 0.665500961sec
{"file_size"=>64000000, "run_parser_called"=>62, "memory"=>"111MB"}: 1.365156626sec
{"file_size"=>128000000, "run_parser_called"=>123, "memory"=>"212MB"}: 3.367049501sec
{"file_size"=>256000000, "run_parser_called"=>245, "memory"=>"413MB"}: 10.04248018sec
{"file_size"=>512000000, "run_parser_called"=>489, "memory"=>"816MB"}: 36.190956998sec
#after
{"file_size"=>1000000, "run_parser_called"=>1, "memory"=>"12MB"}: 0.176965069sec
{"file_size"=>2000000, "run_parser_called"=>2, "memory"=>"12MB"}: 0.176258459sec
{"file_size"=>4000000, "run_parser_called"=>4, "memory"=>"12MB"}: 0.154743461sec
{"file_size"=>8000000, "run_parser_called"=>8, "memory"=>"12MB"}: 0.207943249sec
{"file_size"=>16000000, "run_parser_called"=>16, "memory"=>"12MB"}: 0.320023872sec
{"file_size"=>32000000, "run_parser_called"=>31, "memory"=>"12MB"}: 0.655531422sec
{"file_size"=>64000000, "run_parser_called"=>62, "memory"=>"12MB"}: 1.554511258sec
{"file_size"=>128000000, "run_parser_called"=>123, "memory"=>"12MB"}: 2.579862891sec
{"file_size"=>256000000, "run_parser_called"=>245, "memory"=>"14MB"}: 8.108559264sec
{"file_size"=>512000000, "run_parser_called"=>489, "memory"=>"12MB"}: 37.029166761sec
```